+++ src/actions/login-modal-actions.js
@@ -1,0 +1,8 @@
+import Actions from '../reducers/actions';
+
+export const closeLoginModal = () => ({
+  type: Actions.login.closeLoginModal,
+});
+export const swapModal = () => ({
+  type: Actions.login.swapModal,
+});
