+++ src/actions/navbar-actions.js
@@ -1,0 +1,7 @@
+import Actions from '../reducers/actions';
+
+export const setLoggedIn = () => ({
+  type: Actions.login.isLoggedIn,
+});
+
+export default setLoggedIn;
