+++ src/actions/signup-modal-actions.js
@@ -1,0 +1,8 @@
+import Actions from '../reducers/actions';
+
+export const closeSignupModal = () => ({
+  type: Actions.login.closeSignupModal,
+});
+export const swapModal = () => ({
+  type: Actions.login.swapModal,
+});
