+++ src/containers/connected-pipeline/connected-request-information/connected-request-information.js
@@ -1,5 +1,6 @@
 import { connect } from 'react-redux';
-import Actions from '../../../reducers/actions';
+import { bindActionCreators } from 'redux';
+import * as PipelineActions from '../../../actions/pipeline-actions';
 import requestInformationComponent from '../../../pages/pipeline/request-information';
 
 const mapStateToProps = state => ({
@@ -7,22 +8,13 @@ const mapStateToProps = state => ({
   pageInProgress: state.requests.requestInProgress,
 });
 
-const mapDispatchToProps = dispatch => ({
-  setRequestInformation: (data) => {
-    const action = {
-      type: Actions.requests.setRequestInformation,
-      data,
-    };
-    dispatch(action);
-  },
-  requestInProgress: (data) => {
-    const action = {
-      type: Actions.requests.requestInProgress,
-      data,
-    };
-    dispatch(action);
-  },
-});
+const mapDispatchToProps = dispatch =>
+  bindActionCreators(
+    {
+      ...PipelineActions,
+    },
+    dispatch,
+  );
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
