+++ src/containers/connected-pipeline/connected-request-price/connected-request-price.js
@@ -1,48 +1,20 @@
 import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
+import * as PipelineActions from '../../../actions/pipeline-actions';
 import requestPriceComponent from '../../../pages/pipeline/request-price';
-import Actions from '../../../reducers/actions';
 
 const mapStateToProps = state => ({
   typeOfRequest: state.requests.selection,
   requestPrice: state.requests.requestPrice,
   pageInProgress: state.requests.requestInProgress,
 });
-const mapDispatchToProps = dispatch => ({
-  setRequestPrice: (event) => {
-    const stringValue = event.target.value;
-    let value = parseFloat(stringValue);
-    if (isNaN(value) || value === 0) {
-      value = '';
-    }
-    const action = {
-      type: Actions.requests.setRequestPrice,
-      value,
-    };
-
-    dispatch(action);
-  },
-  incrementRequestPrice: () => {
-    const action = {
-      type: Actions.requests.incrementPrice,
-    };
-
-    dispatch(action);
-  },
-  decrementRequestPrice: () => {
-    const action = {
-      type: Actions.requests.decrementPrice,
-    };
-
-    dispatch(action);
-  },
-  requestInProgress: (data) => {
-    const action = {
-      type: Actions.requests.requestInProgress,
-      data,
-    };
-    dispatch(action);
-  },
-});
+const mapDispatchToProps = dispatch =>
+  bindActionCreators(
+    {
+      ...PipelineActions,
+    },
+    dispatch,
+  );
 
 export default connect(
   mapStateToProps,
