+++ src/containers/connected-pipeline/connected-request-price/connected-request-price.spec.js
@@ -22,12 +22,7 @@ describe('Connected RequestInformation Page', () => {
   it('has the correct props', () => {
     const wrapper = shallow(<RequestPrice store={store} />);
 
-    const dispatchProps = [
-      'setRequestPrice',
-      'decrementRequestPrice',
-      'incrementRequestPrice',
-      'requestInProgress',
-    ];
+    const dispatchProps = ['setRequestPrice', 'requestInProgress'];
 
     dispatchProps.forEach((prop) => {
       expect(wrapper.props()[prop]).toEqual(expect.any(Function));
