+++ src/containers/connected-pipeline/connected-request-selection/connected-request-selection.js
@@ -1,27 +1,19 @@
 import { connect } from 'react-redux';
-import Actions from '../../../reducers/actions';
+import { bindActionCreators } from 'redux';
+import * as PipelineActions from '../../../actions/pipeline-actions';
 import requestSelectionComponent from '../../../pages/pipeline/request-selection';
 
 const mapStateToProps = state => ({
   pageInProgress: state.requests.requestInProgress,
 });
 
-const mapDispatchToProps = dispatch => ({
-  setTypeOfRequest: (data) => {
-    const action = {
-      type: Actions.requests.setTypeOfRequest,
-      data,
-    };
-    dispatch(action);
-  },
-  requestInProgress: (data) => {
-    const action = {
-      type: Actions.requests.requestInProgress,
-      data,
-    };
-    dispatch(action);
-  },
-});
+const mapDispatchToProps = dispatch =>
+  bindActionCreators(
+    {
+      ...PipelineActions,
+    },
+    dispatch,
+  );
 
 export default connect(
   () => mapStateToProps,
