+++ src/containers/connected-signup-modal/connected-signup-modal.js
@@ -1,26 +1,18 @@
 import { connect } from 'react-redux';
+import { bindActionCreators } from 'redux';
 import signupModalComponent from '../../components/signup-modal';
-import Actions from '../../reducers/actions';
+import * as SignupModalActions from '../../actions/signup-modal-actions';
 
 const mapStateToProps = state => ({
   isSignupModalOpen: state.login.isSignupModalOpen,
 });
-const mapDispatchToProps = dispatch => ({
-  closeSignupModal: () => {
-    const action = {
-      type: Actions.login.closeSignupModal,
-    };
-
-    dispatch(action);
-  },
-  swapModal: () => {
-    const action = {
-      type: Actions.login.swapModal,
-    };
-
-    dispatch(action);
-  },
-});
+const mapDispatchToProps = dispatch =>
+  bindActionCreators(
+    {
+      ...SignupModalActions,
+    },
+    dispatch,
+  );
 
 export default connect(
   mapStateToProps,
